Refactor analyze whitelist into own file
